refactor: move repeated code from constants to mt5AccountIcon
