+++ packages/tradershub/src/features/cfd/flows/MT5/AvailableMT5AccountsList/AvailableMT5AccountsList.tsx
@@ -1,32 +1,11 @@
-import React, { FC } from 'react';
+import React from 'react';
 import { Button, Text } from '@deriv/quill-design';
 import { TradingAccountCard } from '../../../../../components';
-import { getStaticUrl } from '../../../../../helpers/urls';
 import { THooks } from '../../../../../types';
 import { MarketTypeDetails } from '../../../constants';
+import { MT5AccountIcon } from '../MT5AccountIcon';
 
-type TProps = {
-    account: THooks.SortedMT5Accounts;
-};
-
-const MT5AccountIcon: FC<TProps> = ({ account }) => {
-    const handleClick = () => {
-        window.open(getStaticUrl('/dmt5'));
-    };
-    const handleKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
-        // Fix sonarcloud issue
-        if (event.key === 'Enter' || event.key === ' ') {
-            handleClick();
-        }
-    };
-    return (
-        <div className='cursor-pointer' onClick={handleClick} onKeyDown={handleKeyDown} role='button' tabIndex={0}>
-            {MarketTypeDetails[account.market_type || 'all'].icon}
-        </div>
-    );
-};
-
-const AvailableMT5AccountsList: FC<TProps> = ({ account }) => {
+const AvailableMT5AccountsList = ({ account }: { account: THooks.MT5AccountsList }) => {
     const { description, title } = MarketTypeDetails[account.market_type || 'all'];
 
     return (
